+++ thunx-autoconfigure/src/main/java/com/contentgrid/thunx/gateway/autoconfigure/GatewayAutoConfiguration.java
@@ -1,7 +1,8 @@
 package com.contentgrid.thunx.gateway.autoconfigure;
 
 import com.contentgrid.opa.client.OpaClient;
-import com.contentgrid.opa.client.rest.RestClientConfiguration;
+import com.contentgrid.thunx.opa.autoconfigure.OpaClientAutoConfiguration;
+import com.contentgrid.thunx.opa.autoconfigure.OpaProperties;
 import com.contentgrid.thunx.pdp.PolicyDecisionComponentImpl;
 import com.contentgrid.thunx.pdp.PolicyDecisionPointClient;
 import com.contentgrid.thunx.pdp.opa.OpaInputProvider;
@@ -11,10 +12,10 @@ import com.contentgrid.thunx.spring.gateway.filter.AbacGatewayFilterFactory;
 import com.contentgrid.thunx.spring.security.DefaultOpaInputProvider;
 import com.contentgrid.thunx.spring.security.ReactivePolicyAuthorizationManager;
 import org.springframework.boot.autoconfigure.AutoConfiguration;
+import org.springframework.boot.autoconfigure.AutoConfigureAfter;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnBean;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnClass;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnWebApplication;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnWebApplication.Type;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
@@ -26,20 +27,11 @@ import org.springframework.security.web.server.authorization.AuthorizationContex
 import org.springframework.web.server.ServerWebExchange;
 
 @AutoConfiguration
+@AutoConfigureAfter(OpaClientAutoConfiguration.class)
 @ConditionalOnClass({OpaClient.class, AbstractGatewayFilterFactory.class})
 @EnableConfigurationProperties(OpaProperties.class)
 @ConditionalOnWebApplication(type = Type.REACTIVE)
 public class GatewayAutoConfiguration {
-
-    @Bean
-    @ConditionalOnMissingBean
-    @ConditionalOnProperty("opa.service.url")
-    public OpaClient opaClient(OpaProperties opaProperties) {
-        return OpaClient.builder()
-                .httpLogging(RestClientConfiguration.LogSpecification::all)
-                .url(opaProperties.getService().getUrl())
-                .build();
-    }
 
     @Bean
     @ConditionalOnMissingBean