+++ thunx-autoconfigure/src/test/java/com/contentgrid/thunx/api/autoconfigure/AbacAutoConfigurationTest.java
@@ -4,7 +4,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 import com.contentgrid.thunx.encoding.ThunkExpressionDecoder;
 import com.contentgrid.thunx.gateway.autoconfigure.GatewayAutoConfiguration;
-import com.contentgrid.thunx.gateway.autoconfigure.OpaProperties;
+import com.contentgrid.thunx.spring.gateway.filter.AbacGatewayFilterFactory;
 import com.contentgrid.thunx.spring.security.AbacRequestFilter;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.UnsatisfiedDependencyException;
@@ -17,7 +17,7 @@ import org.springframework.boot.test.context.runner.WebApplicationContextRunner;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.data.querydsl.binding.QuerydslBindingsFactory;
 
-public class AbacAutoConfigurationTest {
+class AbacAutoConfigurationTest {
 
     WebApplicationContextRunner contextRunner = new WebApplicationContextRunner()
             .withPropertyValues("spring.cloud.gateway.server.webflux.enabled=false")
@@ -27,7 +27,7 @@ public class AbacAutoConfigurationTest {
             ));
 
     @Test
-    public void shouldEnableAbacByDefault() {
+    void shouldEnableAbacByDefault() {
 
         contextRunner.withUserConfiguration(TestContext.class)
                 .run((context) -> {
@@ -43,7 +43,7 @@ public class AbacAutoConfigurationTest {
     }
 
     @Test
-    public void shouldEnableAbacWhenPropertyEqualsHeader() {
+    void shouldEnableAbacWhenPropertyEqualsHeader() {
 
         contextRunner.withUserConfiguration(TestContext.class)
                 .withSystemProperties("contentgrid.thunx.abac.source=header")
@@ -60,7 +60,7 @@ public class AbacAutoConfigurationTest {
     }
 
     @Test
-    public void shouldDisableAbacWhenPropertyEqualsNone() {
+    void shouldDisableAbacWhenPropertyEqualsNone() {
 
         contextRunner.withUserConfiguration(TestContext.class)
                 .withSystemProperties("contentgrid.thunx.abac.source=none")
@@ -77,7 +77,7 @@ public class AbacAutoConfigurationTest {
     }
 
     @Test
-    public void shouldFailWhenPropertyIsInvalid() {
+    void shouldFailWhenPropertyIsInvalid() {
         contextRunner.withUserConfiguration(TestContext.class)
                 .withSystemProperties("contentgrid.thunx.abac.source=invalid")
                 .run(context -> {
@@ -99,7 +99,7 @@ public class AbacAutoConfigurationTest {
                     assertThat(context).hasNotFailed();
                     assertThat(context).hasSingleBean(AbacRequestFilter.class);
 
-                    assertThat(context).doesNotHaveBean(OpaProperties.class);
+                    assertThat(context).doesNotHaveBean(AbacGatewayFilterFactory.class);
                 });
     }
 
@@ -117,8 +117,7 @@ public class AbacAutoConfigurationTest {
                     assertThat(context).hasSingleBean(AbacRequestFilter.class);
                     assertThat(context).hasBean("headerAbacContextSupplier");
 
-                    // Bean for gateway
-                    assertThat(context).doesNotHaveBean(OpaProperties.class);
+                    assertThat(context).doesNotHaveBean(AbacGatewayFilterFactory.class);
                     // Beans for thunx-predicates-querydsl (rely on spring-data-querydsl)
                     assertThat(context).doesNotHaveBean(QuerydslBindingsFactory.class);
                     assertThat(context).doesNotHaveBean("interceptRepositoryRestMvcConfiguration");