+++ thunx-autoconfigure/src/test/java/com/contentgrid/thunx/gateway/autoconfigure/GatewayAutoConfigurationTest.java
@@ -7,6 +7,8 @@ import com.contentgrid.opa.client.OpaClient;
 import com.contentgrid.thunx.api.autoconfigure.AbacAutoConfiguration;
 import com.contentgrid.thunx.api.autoconfigure.AbacContextAutoConfiguration;
 import com.contentgrid.thunx.api.autoconfigure.JwtAbacAutoConfiguration;
+import com.contentgrid.thunx.opa.autoconfigure.OpaClientAutoConfiguration;
+import com.contentgrid.thunx.opa.autoconfigure.OpaProperties;
 import com.contentgrid.thunx.pdp.PolicyDecisionPointClient;
 import com.contentgrid.thunx.pdp.opa.OpaQueryProvider;
 import com.contentgrid.thunx.spring.gateway.filter.AbacGatewayFilterFactory;
@@ -27,10 +29,11 @@ import org.springframework.security.core.Authentication;
 import org.springframework.security.web.server.authorization.AuthorizationContext;
 import org.springframework.web.server.ServerWebExchange;
 
-public class GatewayAutoConfigurationTest {
+class GatewayAutoConfigurationTest {
 
     ReactiveWebApplicationContextRunner contextRunner = new ReactiveWebApplicationContextRunner()
             .withConfiguration(AutoConfigurations.of(
+                    OpaClientAutoConfiguration.class,
                     GatewayAutoConfiguration.class
             ));